# docs/swagger.yaml
basePath: /
definitions:
  model.ProductSwagger:
    properties:
      amount:
        example: 10
        type: integer
      description:
        example: A high-quality product
        type: string
      title:
        example: Product A
        type: string
    type: object
  model.UserDeleteSwagger:
    properties:
      password:
        example: "12345678"
        type: string
    type: object
  model.UserInputSwagger:
    properties:
      identity:
        example: Alfin Afifi
        type: string
      password:
        example: "12345678"
        type: string
    type: object
  model.UserSwagger:
    properties:
      email:
        example: pakvincent@gmail.com
        type: string
      names:
        example: Labibuddin
        type: string
      password:
        example: "12345678"
        type: string
      username:
        example: Alfin Afifi
        type: string
    type: object
  model.UserUpdateSwagger:
    properties:
      names:
        example: Udin Labib
        type: string
    type: object
host: localhost:3000
info:
  contact:
    email: labibuddinalfin@gmail.com
    name: Alfin Afifi
  description: This is a simple e-Commerce with Auth and postgres as database
  license:
    name: labibuddin 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Simple e-Commerce
  version: "1.0"
paths:
  /api:
    get:
      consumes:
      - application/json
      description: Simple hello endpoint to test the API
      produces:
      - application/json
      responses:
        "200":
          description: Hello World
          schema:
            type: string
      summary: Hello endpoint
      tags:
      - Hello
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate a user and return a JWT token
      parameters:
      - description: User credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/model.UserInputSwagger'
      produces:
      - application/json
      responses:
        "200":
          description: JWT token
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: User login
      tags:
      - Auth
  /api/product:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all products available in the store
      produces:
      - application/json
      responses:
        "200":
          description: List of products
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get all products
      tags:
      - Products
    post:
      consumes:
      - application/json
      description: Create a new product with the provided title, description, and
        amount
      parameters:
      - description: Product data
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/model.ProductSwagger'
      produces:
      - application/json
      responses:
        "200":
          description: Product created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid input
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a new product
      tags:
      - Products
  /api/product/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a specific product by its ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Product deleted
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Product not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete a product by ID
      tags:
      - Products
    get:
      consumes:
      - application/json
      description: Retrieve details of a specific product by its ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Product details
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Product not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get a product by ID
      tags:
      - Products
  /api/user:
    post:
      consumes:
      - application/json
      description: Create a new user with the provided username and email
      parameters:
      - description: User info
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.UserSwagger'
      produces:
      - application/json
      responses:
        "200":
          description: 'status: success, message: Created user, data: {model.User.Username,
            model.User.Email}'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 'status: error, message: Review your input or Couldn''t create
            user'
          schema:
            additionalProperties: true
            type: object
      summary: Create User
      tags:
      - User
  /api/user/{id}:
    delete:
      description: Delete a user based on the provided ID and password for verification
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Password input for user verification
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.UserDeleteSwagger'
      produces:
      - application/json
      responses:
        "200":
          description: User successfully deleted
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid input
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete a user by ID
      tags:
      - User
    get:
      description: Retrieve user details by user ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User found
          schema:
            additionalProperties: true
            type: object
        "404":
          description: No user found with ID
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get a user by ID
      tags:
      - User
    patch:
      description: Update a user's names based on the provided ID and input data
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Update user input
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.UserUpdateSwagger'
      produces:
      - application/json
      responses:
        "200":
          description: User successfully updated
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid input
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update a user's information
      tags:
      - User
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"